Gamestore revision 1
Gamestore offer history is in the LUA module yet due to a numeric type-related bug when sending the message.

+++ path_10_9/src/gamestore.h
@@ -113,7 +113,7 @@ struct StoreCategory{
 
 class GameStore {
     public:
-        static uint8_t HISTORY_ENTRIES_PER_PAGE;
+        static uint16_t HISTORY_ENTRIES_PER_PAGE;
         static void startup(){
             HISTORY_ENTRIES_PER_PAGE=16;
         }
@@ -147,7 +147,7 @@ class GameStore {
 struct HistoryStoreOffer {
     uint32_t time;
     uint8_t mode;
-    int32_t amount;
+    uint32_t amount;
     std::string description;
 };
